+++ leia-processor/src/test/java/com/grookage/leia/mux/resolver/TagBasedNameResolverTest.java
@@ -29,7 +29,7 @@ class TagBasedNameResolverTest {
     @Test
     @SneakyThrows
     void testNameResolver() {
-        final var resolver = new TagBasedNameResolver(() -> List.of("BACKEND1"));
+        final var resolver = new TagBasedNameResolver(() -> List.of("Backend1"));
         final var leiaMessage = ResourceHelper.getResource("mux/leiaMessage.json", LeiaMessage.class);
         Assertions.assertNotNull(leiaMessage);
         final var eligibleBackends = resolver.getEligibleBackends(leiaMessage);